+++ src/AssetBundles/GulpBundleAssetsBundles.php
@@ -1,5 +1,6 @@
 <?php
-/**
+
+/*
  * UserFrosting Assets (http://www.userfrosting.com)
  *
  * @link      https://github.com/userfrosting/assets
@@ -29,14 +30,14 @@ abstract class GulpBundleAssetsBundles implements AssetBundlesInterface
     protected $jsBundles;
 
     /**
-     * Constructor
+     * Constructor.
      *
      * @param string $filePath Path to gulp-bundle-assets file.
      */
     public function __construct($filePath)
     {
         if (!is_string($filePath)) {
-            throw new \InvalidArgumentException('$filePath must of type string but was ' . gettype($filePath));
+            throw new \InvalidArgumentException('$filePath must of type string but was '.gettype($filePath)); // @codeCoverageIgnore
         }
 
         // Initalise bundles.
@@ -71,11 +72,13 @@ abstract class GulpBundleAssetsBundles implements AssetBundlesInterface
     /**
      * Attempts to read the schema file from provided path.
      *
-     * @param  string                $path          Path to schema file.
-     * @param  bool                  $useRepository
+     * @param string $path          Path to schema file.
+     * @param bool   $useRepository Deprecated, will be removed in a future release. Setting this to true is recommended for an easier migration.
+     *
      * @throws FileNotFoundException if file cannot be found.
      * @throws JsonException         if file cannot be parsed as JSON.
-     * @return mixed|Repository      Returns file contents parsed by json_decode or a Repository if $useRepository is true.
+     *
+     * @return mixed|Repository Returns file contents parsed by json_decode or a Repository if $useRepository is true.
      */
     protected function readSchema($path, $useRepository = false)
     {
@@ -84,13 +87,18 @@ abstract class GulpBundleAssetsBundles implements AssetBundlesInterface
             try {
                 $loader = new YamlFileLoader($path);
 
-                return new Repository($loader->load());
+                return new Repository($loader->load(false));
             } catch (FileNotFoundException $e) {
                 throw new FileNotFoundException('The schema file could not be found.', 0, $e);
             } catch (JsonException $e) {
                 throw new JsonException('The schema file could not be found.', 0, $e);
             }
         } else {
+            // @codeCoverageIgnoreStart
+            // This code path exists to facilitate an easier migration to v5 of this package
+            // It cannot be properly tested here without increasing technical debt substantially and such such it will be removed in the future
+            trigger_error('Usage of \'readSchema\' with repository result disabled is deprecated and will be removed in future releases.', E_USER_DEPRECATED);
+
             // Read schema without abstractions
             if (!file_exists($path)) {
                 throw new FileNotFoundException("The schema '$path' could not be found.");
@@ -103,10 +111,11 @@ abstract class GulpBundleAssetsBundles implements AssetBundlesInterface
 
             $schema = json_decode($doc);
             if ($schema === null) {
-                throw new JsonException("The schema '$path' does not contain a valid JSON document.  JSON error: " . json_last_error());
+                throw new JsonException("The schema '$path' does not contain a valid JSON document.  JSON error: ".json_last_error());
             }
 
             return $schema;
+            // @codeCoverageIgnoreEnd
         }
     }
 }