+++ src/AssetBundles/GulpBundleAssetsCompiledBundles.php
@@ -1,5 +1,6 @@
 <?php
-/**
+
+/*
  * UserFrosting Assets (http://www.userfrosting.com)
  *
  * @link      https://github.com/userfrosting/assets
@@ -9,9 +10,9 @@
 
 namespace UserFrosting\Assets\AssetBundles;
 
+use UserFrosting\Assets\Exception\InvalidBundlesFileException;
 use UserFrosting\Support\Exception\FileNotFoundException;
 use UserFrosting\Support\Exception\JsonException;
-use UserFrosting\Assets\Exception\InvalidBundlesFileException;
 
 /**
  * Represents a collection of asset bundles, loaded from a gulp-bundle-assets results file.
@@ -23,6 +24,7 @@ class GulpBundleAssetsCompiledBundles extends GulpBundleAssetsBundles
 {
     /**
      * {@inheritdoc}
+     *
      * @throws FileNotFoundException       if file cannot be found.
      * @throws JsonException               if file cannot be parsed as JSON.
      * @throws InvalidBundlesFileException if unexpected value encountered.
@@ -40,14 +42,14 @@ class GulpBundleAssetsCompiledBundles extends GulpBundleAssetsBundles
             if (is_string($styles)) {
                 $this->cssBundles[$bundleName][] = $styles;
             } elseif ($styles !== null) {
-                throw new InvalidBundlesFileException("Expected styles property for '$bundleName' to be of type string but was '" . gettype($styles) . "' for '$path'");
+                throw new InvalidBundlesFileException("Expected styles property for '$bundleName' to be of type string but was '".gettype($styles)."' for '$path'");
             }
 
             $scripts = $schema["$bundleName.scripts"];
             if (is_string($scripts)) {
                 $this->jsBundles[$bundleName][] = $scripts;
             } elseif ($scripts !== null) {
-                throw new InvalidBundlesFileException("Expected scripts property for '$bundleName' to be of type string but was '" . gettype($scripts) . "' for '$path'");
+                throw new InvalidBundlesFileException("Expected scripts property for '$bundleName' to be of type string but was '".gettype($scripts)."' for '$path'");
             }
         }
     }