+++ tests/AssetsTemplatePluginTest.php
@@ -1,10 +1,18 @@
 <?php
 
+/*
+ * UserFrosting Assets (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/assets
+ * @copyright Copyright (c) 2013-2019 Alexander Weissman, Jordan Mele
+ * @license   https://github.com/userfrosting/assets/blob/master/LICENSE.md (MIT License)
+ */
+
 use PHPUnit\Framework\TestCase;
-use UserFrosting\UniformResourceLocator\ResourceLocator;
-use UserFrosting\Assets\Assets;
 use UserFrosting\Assets\AssetBundles\GulpBundleAssetsRawBundles;
+use UserFrosting\Assets\Assets;
 use UserFrosting\Assets\AssetsTemplatePlugin;
+use UserFrosting\UniformResourceLocator\ResourceLocator;
 
 /**
  * Tests AssetsForTemplates class.
@@ -28,16 +36,16 @@ class AssetsTemplatePluginTest extends TestCase
      */
     public function setUp()
     {
-        $this->basePath = __DIR__ . '/data';
-        $this->baseUrl = "https://assets.userfrosting.com/assets/";
-        $this->locatorScheme = "assets";
+        $this->basePath = __DIR__.'/data';
+        $this->baseUrl = 'https://assets.userfrosting.com/assets/';
+        $this->locatorScheme = 'assets';
         $this->locator = new ResourceLocator($this->basePath);
         $this->locator->registerStream($this->locatorScheme, '', 'assets');
         $this->locator->registerStream($this->locatorScheme, 'vendor', 'assets', true);
         $this->locator->registerLocation('hawks', 'sprinkles/hawks/');
         $this->locator->registerLocation('owls', 'sprinkles/owls/');
         $this->assets = new Assets($this->locator, $this->locatorScheme, $this->baseUrl);
-        $this->assets->addAssetBundles(new GulpBundleAssetsRawBundles(__DIR__ . "/data/bundle.config.json"));
+        $this->assets->addAssetBundles(new GulpBundleAssetsRawBundles(__DIR__.'/data/bundle.config.json'));
     }
 
     /**
@@ -50,6 +58,7 @@ class AssetsTemplatePluginTest extends TestCase
     {
         $plugin = new AssetsTemplatePlugin($this->assets);
         $this->assertInstanceOf(AssetsTemplatePlugin::class, $plugin);
+
         return $plugin;
     }
 
@@ -57,6 +66,7 @@ class AssetsTemplatePluginTest extends TestCase
      * Test JS bundle method.
      *
      * @param AssetsTemplatePlugin $plugin
+     *
      * @return void
      *
      * @depends testConstructAssetsTemplatePlugin
@@ -70,6 +80,7 @@ class AssetsTemplatePluginTest extends TestCase
      * Test JS bundle method with attributes.
      *
      * @param AssetsTemplatePlugin $plugin
+     *
      * @return void
      *
      * @depends testConstructAssetsTemplatePlugin
@@ -83,6 +94,7 @@ class AssetsTemplatePluginTest extends TestCase
      * Test CSS bundle method.
      *
      * @param AssetsTemplatePlugin $plugin
+     *
      * @return void
      *
      * @depends testConstructAssetsTemplatePlugin
@@ -96,25 +108,27 @@ class AssetsTemplatePluginTest extends TestCase
      * Test CSS bundle method with attributes.
      *
      * @param AssetsTemplatePlugin $plugin
+     *
      * @return void
      *
      * @depends testConstructAssetsTemplatePlugin
      */
     public function testCssBundleWithAttributes(AssetsTemplatePlugin $plugin)
     {
-        $this->assertEquals($plugin->css('test', ['want-async', 'data-test' => "value"]), '<link href="https://assets.userfrosting.com/assets/vendor/bootstrap/css/bootstrap.css" rel="stylesheet" type="text/css" want-async data-test="value" />');
+        $this->assertEquals($plugin->css('test', ['want-async', 'data-test' => 'value']), '<link href="https://assets.userfrosting.com/assets/vendor/bootstrap/css/bootstrap.css" rel="stylesheet" type="text/css" want-async data-test="value" />');
     }
 
     /**
      * Test url method.
      *
      * @param AssetsTemplatePlugin $plugin
+     *
      * @return void
      *
      * @depends testConstructAssetsTemplatePlugin
      */
     public function testUrl(AssetsTemplatePlugin $plugin)
     {
-        $this->assertEquals($plugin->url("assets://vendor/bootstrap/fonts/glyphicons-halflings-regular.eot"), 'https://assets.userfrosting.com/assets/vendor/bootstrap/fonts/glyphicons-halflings-regular.eot');
+        $this->assertEquals($plugin->url('assets://vendor/bootstrap/fonts/glyphicons-halflings-regular.eot'), 'https://assets.userfrosting.com/assets/vendor/bootstrap/fonts/glyphicons-halflings-regular.eot');
     }
 }